refactor(ToolWidgets): use pointState in Polygon and Ruler
