+++ src/vtk/ToolWidgetUtils/pointState.js
